Renamed implementations of LwjglAudio and LwjglInput to follow naming convention
