+++ backends/gdx-backend-lwjgl/src/com/badlogic/gdx/backends/lwjgl/LwjglApplication.java
@@ -22,7 +22,6 @@ import com.badlogic.gdx.backends.lwjgl.audio.LwjglAudio;
 import org.lwjgl.LWJGLException;
 import org.lwjgl.opengl.Display;
 
-import com.badlogic.gdx.Application;
 import com.badlogic.gdx.ApplicationListener;
 import com.badlogic.gdx.ApplicationLogger;
 import com.badlogic.gdx.Audio;
@@ -32,7 +31,7 @@ import com.badlogic.gdx.Input;
 import com.badlogic.gdx.LifecycleListener;
 import com.badlogic.gdx.Net;
 import com.badlogic.gdx.Preferences;
-import com.badlogic.gdx.backends.lwjgl.audio.OpenALAudio;
+import com.badlogic.gdx.backends.lwjgl.audio.OpenALLwjglAudio;
 import com.badlogic.gdx.utils.Array;
 import com.badlogic.gdx.utils.Clipboard;
 import com.badlogic.gdx.utils.GdxRuntimeException;
@@ -276,13 +275,13 @@ public class LwjglApplication implements LwjglApplicationBase {
 
 	@Override
 	public LwjglAudio createAudio(LwjglApplicationConfiguration config) {
-		return new OpenALAudio(config.audioDeviceSimultaneousSources, config.audioDeviceBufferCount,
+		return new OpenALLwjglAudio(config.audioDeviceSimultaneousSources, config.audioDeviceBufferCount,
 			config.audioDeviceBufferSize);
 	}
 
 	@Override
 	public LwjglInput createInput(LwjglApplicationConfiguration config) {
-		return new LwjglInputImpl();
+		return new DefaultLwjglInput();
 	}
 
 	@Override
